Fixing type in MarkerLineNumberView.

+++ ManDrake/LineNumbering/MarkerLineNumberView.m
@@ -96,7 +96,7 @@
 - (void)mouseDown:(NSEvent *)theEvent
 {
 	NSPoint					location;
-	unsigned				line;
+	NSUInteger				line;
 	
 	location = [self convertPoint:[theEvent locationInWindow] fromView:nil];
 	line = [self lineNumberForLocation:location.y];
